Refactor portal content values architecture.
