+++ ethportal-peertest/src/scenarios/offer_accept.rs
@@ -3,11 +3,11 @@ use std::str::FromStr;
 use serde_json::json;
 use tracing::info;
 
-use ethportal_api::types::{content_value::PossibleHistoryContentValue, enr::Enr};
+use ethportal_api::types::enr::Enr;
 use ethportal_api::utils::bytes::hex_encode;
 use ethportal_api::{
     jsonrpsee::async_client::Client, HistoryContentKey, HistoryContentValue,
-    HistoryNetworkApiClient,
+    HistoryNetworkApiClient, PossibleHistoryContentValue,
 };
 
 use crate::{
