+++ rpc/src/history_rpc.rs
@@ -2,7 +2,6 @@ use crate::jsonrpsee::core::{async_trait, RpcResult};
 use anyhow::anyhow;
 use discv5::enr::NodeId;
 use ethportal_api::types::constants::CONTENT_ABSENT;
-use ethportal_api::types::content_value::PossibleHistoryContentValue;
 use ethportal_api::types::enr::Enr;
 use ethportal_api::types::jsonrpc::endpoints::HistoryEndpoint;
 use ethportal_api::types::jsonrpc::request::HistoryJsonRpcRequest;
@@ -13,6 +12,7 @@ use ethportal_api::types::portal::{
 use ethportal_api::HistoryContentKey;
 use ethportal_api::HistoryContentValue;
 use ethportal_api::HistoryNetworkApiServer;
+use ethportal_api::PossibleHistoryContentValue;
 use ethportal_api::RoutingTableInfo;
 use serde_json::{from_value, Value};
 use tokio::sync::mpsc;
