+++ trin-beacon/src/jsonrpc.rs
@@ -1,12 +1,13 @@
 use std::sync::Arc;
 
 use discv5::enr::NodeId;
-use ethportal_api::types::content_value::{BeaconContentValue, ContentValue};
+use ethportal_api::types::content_value::ContentValue;
 use ethportal_api::types::jsonrpc::endpoints::BeaconEndpoint;
 use ethportal_api::types::jsonrpc::request::BeaconJsonRpcRequest;
 use ethportal_api::types::portal::{AcceptInfo, FindNodesInfo, PongInfo, TraceContentInfo};
 use ethportal_api::types::{constants::CONTENT_ABSENT, query_trace::QueryTrace};
 use ethportal_api::utils::bytes::hex_encode;
+use ethportal_api::BeaconContentValue;
 use ethportal_api::{BeaconContentKey, OverlayContentKey, RawContentKey};
 use portalnet::storage::ContentStore;
 use serde_json::{json, Value};
